+++ code/lib/core/i18n/l10n/app_localizations_de.dart
@@ -771,6 +771,9 @@ class AppLocalizationsDe extends AppLocalizations {
 
   @override
   String get paragraphToParagraph => 'In Absatz umwandeln';
+
+  @override
+  String get paragraphLooseList => 'Lockerer Listeneintrag';
   @override
   String get close => 'Schließen';