+++ code/lib/core/i18n/l10n/app_localizations_fr.dart
@@ -773,6 +773,9 @@ class AppLocalizationsFr extends AppLocalizations {
 
   @override
   String get paragraphToParagraph => 'Convertir en paragraphe';
+
+  @override
+  String get paragraphLooseList => 'Élément de liste espacé';
   @override
   String get close => 'Fermer';