+++ code/lib/core/i18n/l10n/app_localizations_ru.dart
@@ -770,6 +770,9 @@ class AppLocalizationsRu extends AppLocalizations {
 
   @override
   String get paragraphToParagraph => 'Преобразовать в абзац';
+
+  @override
+  String get paragraphLooseList => 'Разреженный список';
   @override
   String get close => 'Закрыть';